added votes

+++ app/controllers/messages_controller.rb
@@ -3,12 +3,49 @@ class MessagesController < ApplicationController
     @message = current_user.messages.create(message_params)
   end
 
-  def like
-    message.increment!(:likes)
+  def vote
+    vote = Vote.where(user_id: current_user.id, message_id: message.id)
+    if vote.count.zero?
+      if params[:like]
+        message.increment!(:likes)
+        message.votes.create(user_id: current_user.id, vote: :like)
+      elsif params[:dislike]
+        message.increment!(:dislikes)
+        message.votes.create(user_id: current_user.id, vote: :dislike)
+      end
+    else
+      change_vote(vote.first, message)
+    end
   end
 
-  def dislike
-    message.increment!(:dislikes)
+  def change_vote(vote, message)
+    if params[:like]
+      case vote.vote
+        when 'like' then
+          vote.update(vote: :no_vote)
+          message.decrement!(:likes)
+        when 'dislike' then
+          vote.update(vote: :like)
+          message.decrement!(:dislikes)
+          message.increment!(:likes)
+        when 'no_vote' then
+          vote.update(vote: :like)
+          message.increment!(:likes)
+      end
+    elsif params[:dislike]
+      case vote.vote
+        when 'like' then
+          vote.update(vote: :dislike)
+          message.decrement!(:likes)
+          message.increment!(:dislikes)
+        when 'dislike' then
+          vote.update(vote: :no_vote)
+          message.decrement!(:dislikes)
+        when 'no_vote' then
+          vote.update(vote: :dislike)
+          message.increment!(:dislikes)
+      end
+    end
   end
 
   def edit
@@ -19,11 +56,7 @@ class MessagesController < ApplicationController
   end
 
   def destroy
-    if is_my_message?(message)
-      message.destroy
-    else
-      redirect_to chat_path,  notice: 'Haha. Nothing you can not do.Cunning Ass'
-    end
+    message.destroy if is_my_message?(message)
   end
 
   private
